requirements.txt
flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
flask-socketio==5.3.6
python-socketio==5.11.1 